third_party/blink/web_tests/external/wpt/html/semantics/embedded-content/the-video-element/intrinsic_sizes-expected.txt

This is a testharness.js-based test.
[FAIL] default height is half the width
  assert_equals: expected "200px" but got "150px"
[FAIL] default width is twice the height
  assert_equals: expected "200px" but got "300px"
Harness: the test ran to completion.
